premali: first pass at actually linking
We now no longer hardcoding varying locations, we rewrite vertex varying
locations based on the fragment shaders ordering.

+++ premali/lib/symbols.h
@@ -41,7 +41,7 @@ struct symbol {
 	int element_count;
 
 	int size; /* size it takes in memory. */
-	int offset; /* offset from start of memory block */
+	short offset; /* offset from start of memory block */
 
 	void *address;
 	int physical; /* empty for uniforms */
